pyproject.toml
[project]
name = "nebius-hackathon"
version = "0.1.0"
description = "Self-expanding robot agent — detects capability gaps, searches hardware catalogs, and installs new skills"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28",
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sse-starlette>=3.3.2",
]

[project.scripts]
demo = "orchestrator:main"
store = "store.app:app"
robot = "robot.app:app"
